requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
oauthlib==2.1.0
pycparser==2.18
PyJWT==1.6.4
PyMySQL==0.9.2
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
